Add styling to the baseball intro section and imgs
I positioned the background image for mobile users, as we are targeting
mobile users first.
Applied flex to the logo-intro division and laid out its child elements
to be centered vertical and horizontally.
Applied a white text color to the content within that division and gave
it a very slight text shadow of black as the white font was not very
readable without it.
Also changed the width of the baseball icon svg to 5em, I may change
this later though.