Add styles for image credit span, adjust img
The credit to the photographer was hard to read or be caught by eyes
when scrolling down the page, changing the font color to something
lighter fixes that. I may add a pseudo style to the anchor element
later for when it is being clicked on.
Also added a touch of margin to the right of the icon image to give
some separation from the text, however no I feel like the icon needs to
be a lighter color either gray or white to make it pop out more to the
user.